# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Zond Explorer",
        "title": "Zond Explorer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/address/{address}/balance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the balance information of an address (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Get address balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AddressResponse"
                        }
                    },
                    "403": {
                        "description": "Admin access required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/address/{address}/details": {
            "get": {
                "description": "Get full details of a specific address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Get address details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AddressDetailResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/address/{address}/transactions": {
            "get": {
                "description": "Get list of transactions related to the given address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Get address transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TransactionDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/search/suggestions": {
            "get": {
                "description": "Return suggestions based on a partial query input (e.g. address, tx hash, block number)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Get search suggestions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search input string (partial hash, address, or block number)",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchSuggestionsResponse"
                        }
                    },
                    "400": {
                        "description": "Query parameter required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/validators": {
            "get": {
                "description": "Retrieve a list of all active validators",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Validators"
                ],
                "summary": "Get validators",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ValidatorsResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch validators",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/validators/{validatorId}": {
            "get": {
                "description": "Retrieve a specific validator by its index",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Validators"
                ],
                "summary": "Get validator by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validator index",
                        "name": "validatorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ValidatorDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid validator index",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Validator not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddressDetailResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "total_gas_used": {
                    "type": "integer"
                },
                "transaction_count": {
                    "type": "integer"
                }
            }
        },
        "dto.AddressResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                }
            }
        },
...
        "dto.SearchSuggestionsResponse": {
            "type": "object",
            "properties": {
                "suggestions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Suggestion"
                    }
                }
            }
        },
        "dto.Suggestion": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "example": "transaction"
                },
                "value": {
                    "type": "string",
                    "example": "0xabc123..."
                }
            }
        },
        "dto.TransactionDTO": {
            "type": "object",
            "properties": {
                "block_number": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto.ValidatorDetailResponse": {
            "type": "object",
            "properties": {
                "activation_epoch": {
                    "type": "integer",
                    "example": 194
                },
                "balance": {
                    "type": "integer",
                    "example": 32000000000
                },
                "chain_id": {
                    "type": "integer",
                    "example": 1
                },
                "effective_balance": {
                    "type": "integer",
                    "example": 32000000000
                },
                "exit_epoch": {
                    "type": "integer",
                    "example": 0
                },
                "index": {
                    "type": "integer",
                    "example": 1234
                },
                "public_key": {
                    "type": "string",
                    "example": "0xabc123..."
                },
                "retrieved_from": {
                    "type": "string",
                    "example": "prysm"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "dto.ValidatorResponse": {
            "type": "object",
            "properties": {
                "activation_epoch": {
                    "type": "integer",
                    "example": 194
                },
                "balance": {
                    "type": "string",
                    "example": "32000000000"
                },
                "chain_id": {
                    "type": "integer",
                    "example": 1
                },
                "effective_balance": {
                    "type": "string",
                    "example": "32000000000"
                },
                "exit_epoch": {
                    "type": "integer",
                    "example": 0
                },
                "index": {
                    "type": "integer",
                    "example": 1234
                },
                "public_key": {
                    "type": "string",
                    "example": "0xabc123..."
                },
                "retrieved_from": {
                    "type": "string",
                    "example": "lighthouse"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "dto.ValidatorsResponse": {
            "type": "object",
            "properties": {
                "validators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ValidatorResponse"
                    }
                }
            }
        }
    }
}